+++ rottenpotato/src/main/java/com/redwood/rottenpotato/main/mvccontollers/TVMvcController.java
@@ -63,7 +63,7 @@ public class TVMvcController {
         if (criticScoreCount == 0) {
             model.addAttribute("criticRating", "N/A");
         } else {
-            model.addAttribute("criticRating", String.format("%.2f", criticScore / criticScoreCount));
+            model.addAttribute("criticRating", String.format("%.2f", (double)criticScore / criticScoreCount));
         }
         List<UserRating> userRatings = userRatingRepository.findByItemKey(tv.getTVKey());
         int userScore = 0;
@@ -77,7 +77,7 @@ public class TVMvcController {
         if (userScoreCount == 0) {
             model.addAttribute("userRating", "N/A");
         } else {
-            model.addAttribute("userRating", String.format("%.2f", userScore / userScoreCount));
+            model.addAttribute("userRating", String.format("%.2f", (double)userScore / userScoreCount));
         }
 
         //TODO poster
